py-omop2neo4j-lpg/pyproject.toml
[tool.poetry]
name = "py-omop2neo4j-lpg"
version = "0.1.0"
description = "A robust, high-performance Python package to orchestrate the migration of OMOP vocabulary tables from PostgreSQL to a mature Labeled Property Graph (LPG) in Neo4j."
authors = ["Gowtham Rao <rao@ohdsi.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "*"
neo4j = "*"
pandas = "*"
numpy = "*"
click = "*"
pydantic-settings = "*"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
py-omop2neo4j-lpg = "omop2neo4j_lpg.cli:main"
